pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.6"
description = "Difference calculator - a cli utility to find the difference between two config files"
authors = ["Aleksey Ivanov <aleksey.vl.ivanov@gmail.com>"]
classifiers = [
  "Environment :: Console",
  "Topic :: Utilities",
]
packages = [{ include = "gendiff" }]

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
cohesion = "^1.0.0"
dlint = "^0.12.0"
flake8 = "^4.0.1"
flake8-builtins = "^1.5.3"
flake8-class-attributes-order = "^0.1.2"
flake8-coding = "^1.3.2"
flake8-mutable = "^1.2.0"
flake8-pytest = "^1.3"
flake8-pytest-style = "^1.6.0"
isort = "^5.10.1"
pytest = "^7.0.1"
pytest-cov = "^3.0.0"
rope = "^0.22.0"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
